s3-upload/requirements.txt
boto3>=1.28.0
PyYAML>=6.0
schedule>=1.2.0
